made choosing contract form functional
