+++ web/components/new-contract/contract-params-form.tsx
@@ -1,22 +1,13 @@
 import { ExternalLinkIcon } from '@heroicons/react/outline'
-import clsx from 'clsx'
 import dayjs from 'dayjs'
 import router from 'next/router'
 import { useEffect, useState } from 'react'
 
-import {
-  MAX_DESCRIPTION_LENGTH,
-  MAX_QUESTION_LENGTH,
-  NON_BETTING_OUTCOMES,
-  OutcomeType,
-  Visibility,
-} from 'common/contract'
-import { UNIQUE_BETTOR_BONUS_AMOUNT, getAnte } from 'common/economy'
+import { MAX_QUESTION_LENGTH, OutcomeType } from 'common/contract'
+import { UNIQUE_BETTOR_BONUS_AMOUNT } from 'common/economy'
 import { ENV_CONFIG } from 'common/envs/constants'
-import { Group, groupPath } from 'common/group'
-import { User } from 'common/user'
+import { groupPath } from 'common/group'
 import { formatMoney } from 'common/util/format'
-import { removeUndefinedProps } from 'common/util/object'
 import { MINUTE_MS } from 'common/util/time'
 import { AddFundsModal } from 'web/components/add-funds-modal'
 import { MultipleChoiceAnswers } from 'web/components/answers/multiple-choice-answers'
@@ -26,60 +17,46 @@ import { Row } from 'web/components/layout/row'
 import { Spacer } from 'web/components/layout/spacer'
 import { Checkbox } from 'web/components/widgets/checkbox'
 import { ChoicesToggleGroup } from 'web/components/widgets/choices-toggle-group'
-import { TextEditor, useTextEditor } from 'web/components/widgets/editor'
+import { TextEditor } from 'web/components/widgets/editor'
 import { ExpandingInput } from 'web/components/widgets/expanding-input'
 import { InfoTooltip } from 'web/components/widgets/info-tooltip'
 import { Input } from 'web/components/widgets/input'
 import ShortToggle from 'web/components/widgets/short-toggle'
-import { createMarket } from 'web/lib/firebase/api'
-import { track } from 'web/lib/service/analytics'
-import { safeLocalStorage } from 'web/lib/util/local'
-import { QfExplainer } from './contract/qf-overview'
+import { QfExplainer } from '../contract/qf-overview'
 
-import { generateJSON } from '@tiptap/core'
-import { Contract } from 'common/contract'
-import { STONK_NO, STONK_YES } from 'common/stonk'
-import { extensions } from 'common/util/parse'
-import { usePersistentLocalState } from 'web/hooks/use-persistent-local-state'
-import { getGroup } from 'web/lib/supabase/group'
-import WaitingForSupabaseButton from './contract/waiting-for-supabase-button'
-import { Col } from './layout/col'
-import { BuyAmountInput } from './widgets/amount-input'
+import { useNewContract } from 'web/hooks/use-new-contract'
+import WaitingForSupabaseButton from '../contract/waiting-for-supabase-button'
+import { Col } from '../layout/col'
+import { BuyAmountInput } from '../widgets/amount-input'
+import {
+  ContractVisibilityType,
+  CreateContractStateType,
+  NewQuestionParams,
+} from './new-contract-panel'
+import { User } from 'common/user'
+import { VisibilityTheme } from 'web/pages/create'
 
-export type NewQuestionParams = {
-  groupId?: string
-  q: string
-  type: string
-  description: string
-  closeTime: string
-  outcomeType: string
-  visibility: string
-  // Params for PSEUDO_NUMERIC outcomeType
-  min?: string
-  max?: string
-  isLogScale?: string
-  initValue?: string
-
-  // Answers encoded as:
-  // a0: string
-  // a1: string
-  // ...etc
-}
-
-// Allow user to create a new contract
-export function NewContractPanel(props: {
+export function ContractParamsForm(props: {
+  outcomeType: OutcomeType
+  setOutcomeType: (outcomeType: OutcomeType) => void
+  setState: (state: CreateContractStateType) => void
   creator: User
-  params?: NewQuestionParams
+  setTheme: (theme: VisibilityTheme) => void
   fromGroup?: boolean
-  className?: string
-  setTheme: (theme: 'private' | 'non-private') => void
+  params?: NewQuestionParams
 }) {
-  const { creator, params, fromGroup, className, setTheme } = props
+  const {
+    outcomeType,
+    setOutcomeType,
+    setState,
+    creator,
+    setTheme,
+    fromGroup,
+    params,
+  } = props
   const {
     question,
     setQuestion,
-    outcomeType,
-    setOutcomeType,
     editor,
     closeDate,
     setCloseDate,
@@ -113,14 +90,10 @@ export function NewContractPanel(props: {
     bountyAmount,
     setBountyAmount,
     newContract,
-  } = useNewContract(creator, params)
-
+  } = useNewContract(creator, params, outcomeType)
   const [bountyError, setBountyError] = useState<string | undefined>(undefined)
-  const [fundsModalOpen, setFundsModalOpen] = useState(false)
-  const [toggleVisibility, setToggleVisibility] = useState<
-    'public' | 'unlisted'
-  >('public')
-
+  const [toggleVisibility, setToggleVisibility] =
+    useState<ContractVisibilityType>('public')
   useEffect(() => {
     if (selectedGroup?.privacyStatus == 'private') {
       setVisibility('private')
@@ -131,27 +104,11 @@ export function NewContractPanel(props: {
     }
   }, [selectedGroup?.privacyStatus, toggleVisibility])
 
+  const [fundsModalOpen, setFundsModalOpen] = useState(false)
+
   return (
-    <div className={clsx(className, 'text-ink-1000')}>
+    <>
       <label className="flex px-1 pt-2 pb-3">Answer type</label>
-      <Row>
-        <ChoicesToggleGroup
-          currentChoice={outcomeType}
-          setChoice={(choice) => {
-            setOutcomeType(choice as OutcomeType)
-          }}
-          choicesMap={{
-            'Yes\xa0/ No': 'BINARY', // non-breaking space
-            'Multiple choice': 'MULTIPLE_CHOICE',
-            'Free response': 'FREE_RESPONSE',
-            // Stock: 'STONK',
-            Numeric: 'PSEUDO_NUMERIC',
-            'Bountied Question': 'BOUNTIED_QUESTION',
-          }}
-          disabled={isSubmitting}
-          className={'col-span-4'}
-        />
-      </Row>
       <Spacer h={6} />
       <Col>
         {outcomeType === 'STONK' ? (
@@ -497,279 +454,9 @@ export function NewContractPanel(props: {
       </Row>
 
       <Spacer h={6} />
-    </div>
+    </>
   )
 }
-
-const useNewContract = (
-  creator: User,
-  params: NewQuestionParams | undefined
-) => {
-  // If params specify content like a question, store it separately in local storage.
-  const paramsKey = params?.q ?? ''
-
-  const [outcomeType, setOutcomeType] = usePersistentLocalState<OutcomeType>(
-    (params?.outcomeType as OutcomeType) ?? 'BINARY',
-    'new-outcome-type' + paramsKey
-  )
-  const [minString, setMinString] = usePersistentLocalState(
-    params?.min ?? '',
-    'new-min' + paramsKey
-  )
-  const [maxString, setMaxString] = usePersistentLocalState(
-    params?.max ?? '',
-    'new-max' + paramsKey
-  )
-  const [isLogScale, setIsLogScale] = usePersistentLocalState<boolean>(
-    !!params?.isLogScale,
-    'new-is-log-scale' + paramsKey
-  )
-
-  const [initialValueString, setInitialValueString] = usePersistentLocalState(
-    params?.initValue,
-    'new-init-value' + paramsKey
-  )
-  const [visibility, setVisibility] = usePersistentLocalState<Visibility>(
-    (params?.visibility as Visibility) ?? 'public',
-    `new-visibility${'-' + params?.groupId ?? ''}`
-  )
-  const [newContract, setNewContract] = useState<Contract | undefined>(
-    undefined
-  )
-
-  const paramAnswers = []
-  let i = 0
-  while (params && (params as any)[`a${i}`]) {
-    paramAnswers.push((params as any)[`a${i}`])
-    i++
-  }
-  // for multiple choice, init to 2 empty answers
-  const [answers, setAnswers] = usePersistentLocalState(
-    paramAnswers.length ? paramAnswers : ['', ''],
-    'new-answers' + paramsKey
-  )
-  console.log('paramAnswers', paramAnswers, 'answers', answers)
-
-  const [question, setQuestion] = usePersistentLocalState(
-    '',
-    'new-question' + paramsKey
-  )
-  useEffect(() => {
-    if (params?.q) setQuestion(params?.q ?? '')
-  }, [params?.q])
-
-  useEffect(() => {
-    if (params?.groupId) {
-      getGroup(params?.groupId).then((group) => {
-        if (group) {
-          setSelectedGroup(group)
-        }
-      })
-    }
-  }, [creator.id, params?.groupId])
-
-  const ante = getAnte(outcomeType, answers.length, visibility === 'private')
-
-  // If params.closeTime is set, extract out the specified date and time
-  // By default, close the question a week from today
-  const weekFromToday = dayjs().add(7, 'day').format('YYYY-MM-DD')
-  const timeInMs = Number(params?.closeTime ?? 0)
-  const initDate = timeInMs
-    ? dayjs(timeInMs).format('YYYY-MM-DD')
-    : weekFromToday
-  const initTime = timeInMs ? dayjs(timeInMs).format('HH:mm') : '23:59'
-
-  const [closeDate, setCloseDate] = usePersistentLocalState<undefined | string>(
-    timeInMs ? initDate : undefined,
-    'now-close-date' + paramsKey
-  )
-  const [closeHoursMinutes, setCloseHoursMinutes] = usePersistentLocalState<
-    string | undefined
-  >(timeInMs ? initTime : undefined, 'now-close-time' + paramsKey)
-
-  const [selectedGroup, setSelectedGroup] = usePersistentLocalState<
-    Group | undefined
-  >(undefined, 'new-selected-group' + paramsKey)
-
-  const [bountyAmount, setBountyAmount] = usePersistentLocalState<
-    number | undefined
-  >(50, 'new-bounty' + paramsKey)
-
-  const closeTime = closeDate
-    ? dayjs(`${closeDate}T${closeHoursMinutes}`).valueOf()
-    : undefined
-
-  const balance = creator.balance || 0
-
-  const min = minString ? parseFloat(minString) : undefined
-  const max = maxString ? parseFloat(maxString) : undefined
-  const initialValue = initialValueString
-    ? parseFloat(initialValueString)
-    : undefined
-
-  useEffect(() => {
-    if (outcomeType === 'STONK' || NON_BETTING_OUTCOMES.includes(outcomeType)) {
-      setCloseDate(dayjs().add(1000, 'year').format('YYYY-MM-DD'))
-      setCloseHoursMinutes('23:59')
-
-      if (outcomeType == 'STONK') {
-        if (editor?.isEmpty) {
-          editor?.commands.setContent(
-            generateJSON(
-              `<div>
-            ${STONK_YES}: good<br/>${STONK_NO}: bad<br/>Question trades based on sentiment & never
-            resolves.
-          </div>`,
-              extensions
-            )
-          )
-        }
-      }
-    }
-  }, [outcomeType])
-
-  const isValidMultipleChoice = answers.every(
-    (answer) => answer.trim().length > 0
-  )
-
-  const isValid =
-    question.length > 0 &&
-    ante !== undefined &&
-    ante !== null &&
-    ante <= balance &&
-    // closeTime must be in the future
-    (closeTime ?? Infinity) > Date.now() &&
-    (outcomeType !== 'PSEUDO_NUMERIC' ||
-      (min !== undefined &&
-        max !== undefined &&
-        initialValue !== undefined &&
-        isFinite(min) &&
-        isFinite(max) &&
-        min < max &&
-        max - min > 0.01 &&
-        min < initialValue &&
-        initialValue < max)) &&
-    (outcomeType !== 'MULTIPLE_CHOICE' || isValidMultipleChoice)
-
-  const [errorText, setErrorText] = useState<string>('')
-  useEffect(() => {
-    setErrorText('')
-  }, [isValid])
-
-  const editor = useTextEditor({
-    key: 'create market' + paramsKey,
-    max: MAX_DESCRIPTION_LENGTH,
-    placeholder: descriptionPlaceholder,
-    defaultValue: params?.description
-      ? JSON.parse(params.description)
-      : undefined,
-  })
-
-  function setCloseDateInDays(days: number) {
-    const newCloseDate = dayjs().add(days, 'day').format('YYYY-MM-DD')
-    setCloseDate(newCloseDate)
-  }
-
-  const resetProperties = () => {
-    editor?.commands.clearContent(true)
-    safeLocalStorage?.removeItem(`text create market`)
-    setQuestion('')
-    setOutcomeType('BINARY')
-    setCloseDate(undefined)
-    setCloseHoursMinutes(undefined)
-    setSelectedGroup(undefined)
-    setVisibility((params?.visibility as Visibility) ?? 'public')
-    setAnswers(['', '', ''])
-    setMinString('')
-    setMaxString('')
-    setInitialValueString('')
-    setIsLogScale(false)
-    setBountyAmount(50)
-  }
-
-  const [isSubmitting, setIsSubmitting] = useState(false)
-
-  async function submit() {
-    if (!isValid) return
-    setIsSubmitting(true)
-    try {
-      const newContract = (await createMarket(
-        removeUndefinedProps({
-          question,
-          outcomeType,
-          description: editor?.getJSON(),
-          initialProb: 50,
-          ante,
-          closeTime,
-          min,
-          max,
-          initialValue,
-          isLogScale,
-          answers,
-          groupId: selectedGroup?.id,
-          visibility,
-          utcOffset: new Date().getTimezoneOffset(),
-          totalBounty: bountyAmount,
-        })
-      )) as Contract
-
-      setNewContract(newContract)
-      resetProperties()
-
-      track('create market', {
-        slug: newContract.slug,
-        selectedGroup: selectedGroup?.id,
-        outcomeType,
-      })
-    } catch (e) {
-      console.error('error creating contract', e)
-      setErrorText((e as any).message || 'Error creating contract')
-      setIsSubmitting(false)
-    }
-  }
-  return {
-    question,
-    setQuestion,
-    outcomeType,
-    setOutcomeType,
-    editor,
-    closeDate,
-    setCloseDate,
-    closeHoursMinutes,
-    setCloseHoursMinutes,
-    setCloseDateInDays,
-    min,
-    minString,
-    setMinString,
-    max,
-    maxString,
-    setMaxString,
-    initialValue,
-    initialValueString,
-    setInitialValueString,
-    isLogScale,
-    setIsLogScale,
-    answers,
-    setAnswers,
-    selectedGroup,
-    setSelectedGroup,
-    visibility,
-    setVisibility,
-    initTime,
-    submit,
-    isValid,
-    isSubmitting,
-    errorText,
-    balance,
-    ante,
-    newContract,
-    bountyAmount,
-    setBountyAmount,
-  }
-}
-
-const descriptionPlaceholder =
-  'Optional. Provide background info and question resolution criteria here.'
 
 // get days from today until the end of this year:
 const daysLeftInTheYear = dayjs().endOf('year').diff(dayjs(), 'day')
